+++ parser/src/keys/linux.rs
@@ -771,11 +771,3 @@ impl From<Btn> for OsCode {
         }
     }
 }
-
-#[derive(Debug, Copy, Clone, PartialEq, Eq)]
-pub enum UnicodeTermination {
-    Enter,
-    Space,
-    SpaceEnter,
-    EnterSpace,
-}
